Moved setPanelBackgrounds() to the Level superclass. Every 'level' now has the same background color.
