+++ src/game/levels/Level2.java
@@ -130,6 +130,8 @@ public class Level2 extends Level {
 
         levelContainer.add(sideMenu, levelConstraints);
 
+        setPanelBackgrounds(levelContainer, new Color(206, 206, 206));
+
         return levelContainer;
     }
 }
